Ids signature to Long? and personalized CrudRepositories

+++ src/main/kotlin/ar/edu/unsam/phm/controller/BookController.kt
@@ -20,13 +20,8 @@ class BookController(
     @GetMapping("/filtered-books")
     fun getFilteredBooks(
         @ModelAttribute criteria: BookSearchCriteria,
-//        @RequestParam(defaultValue = "0") page: Int,
-//        @RequestParam(defaultValue = "6") size: Int,
-//        @RequestParam(defaultValue = "title") sortBy: String,
-//        @RequestParam(defaultValue = "true") ascending: Boolean
+
     ): PageResponse<BookDTO> {
-//        println(criteria.toString())
-//        println("$page, $size, $sortBy, $ascending")
         val direction = if (criteria.ascending) Sort.Direction.ASC else Sort.Direction.DESC
         val pageable = PageRequest.of(criteria.page, criteria.pageSize, Sort.by(direction, criteria.sortBy))
         return bookService.searchBooks(criteria, pageable)
@@ -38,21 +33,21 @@ class BookController(
     }
 
     @PutMapping("/edit-book/{id}")
-    fun editBook(@PathVariable id: Int, @RequestBody bookCreateDTO: BookCreateDTO) {
+    fun editBook(@PathVariable id: Long, @RequestBody bookCreateDTO: BookCreateDTO) {
         bookService.updateBook(id, bookCreateDTO)
     }
 
     @DeleteMapping("/delete-book/{id}")
-    fun deleteBook(@PathVariable id: Int) {
+    fun deleteBook(@PathVariable id: Long) {
         bookService.deleteBook(id)
     }
 
     @GetMapping("/book-detail/{id}")
-    fun getBookById(@PathVariable id: Int) =
+    fun getBookById(@PathVariable id: Long) =
         bookService.getBookById(id).toDTO()
 
     @GetMapping("/book-detail/{id}/bibliokarmas")
-    fun calculateBibliokarmas(@PathVariable id: Int, @RequestParam userId: Int, @RequestParam pickUpDate: LocalDate, @RequestParam dropOffDate: LocalDate): Int =
+    fun calculateBibliokarmas(@PathVariable id: Long, @RequestParam userId: Long, @RequestParam pickUpDate: LocalDate, @RequestParam dropOffDate: LocalDate): Int =
         bookService.recalculateBibliokarmas(id, userId, pickUpDate, dropOffDate)
