+++ src/main/kotlin/ar/edu/unsam/phm/dto/CreateReservationDTO.kt
@@ -4,8 +4,8 @@ import ar.edu.unsam.phm.domain.Reservation
 import java.time.LocalDate
 
 data class CreateReservationDTO(
-    val bookId: Int,
-    val sessionId: Int,
+    val bookId: Long,
+    val sessionId: Long,
     val pickUpDate: LocalDate,
     val dropOffDate: LocalDate
 ){}
